src/service1/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
fastapi==0.52.0
h11==0.9.0
idna==2.9
lorem==0.1.1
netifaces==0.10.9
pydantic==1.4
python-consul==1.1.0
requests==2.23.0
six==1.14.0
starlette==0.13.2
urllib3==1.25.8
uvicorn==0.11.3
websockets==8.1
